XRENDERING-736: Macros info, success, warning and error are only distinguished by colors * Moved the classes from `xwiki-rendering-transformation-icon` * Created an IconBlock class to contain the icons
