+++ xwiki-rendering-api/src/main/resources/META-INF/components.txt
@@ -34,6 +34,7 @@ org.xwiki.rendering.internal.syntax.SyntaxConverter
 org.xwiki.rendering.internal.syntax.DefaultSyntaxRegistry
 org.xwiki.rendering.internal.syntax.SyntaxRegistryListener
 org.xwiki.rendering.internal.transformation.DefaultRenderingContext
+org.xwiki.rendering.internal.transformation.DefaultIconProvider
 org.xwiki.rendering.internal.transformation.DefaultTransformationManager
 org.xwiki.rendering.internal.transformation.RenderingContextStore
 org.xwiki.rendering.internal.util.DefaultErrorBlockGenerator
