+++ xwiki-rendering-transformations/xwiki-rendering-transformation-icon/src/main/resources/META-INF/components.txt
@@ -1,3 +1,2 @@
 org.xwiki.rendering.internal.transformation.icon.IconTransformation
 org.xwiki.rendering.internal.transformation.icon.DefaultIconTransformationConfiguration
-org.xwiki.rendering.internal.transformation.icon.DefaultIconProvider
